BackEnd/requirements.txt
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-simpleui==2025.1.13
mysqlclient==2.2.4
Pillow==10.2.0  # 用于图片处理
python-dotenv==1.0.1 
channels==3.0.5
channels-redis==4.1.0  # 用于生产环境中的Channel Layer
daphne==3.0.2  # ASGI服务器，用于生产环境 